perf(tests): Extract simple backup tests
These are much faster if they don't need Robolectric

Takes 5s off test execution
